backend/requirements-final.txt
# ChatIng - Dependencias instaladas y verificadas
# Generado automáticamente - Diciembre 2024
# Compatible con Python 3.12

# === CORE FRAMEWORK ===
fastapi==0.115.13
uvicorn[standard]==0.34.3
pydantic==2.11.7

# === ENVIRONMENT & CONFIG ===
python-dotenv==1.1.1

# === LLM INTEGRATIONS ===
anthropic==0.55.0
openai==1.91.0
google-generativeai==0.8.5

# === LANGCHAIN ECOSYSTEM ===
langchain==0.3.26
langchain-core==0.3.66
langchain-chroma==0.2.4

# === MEMORY & VECTORIZATION ===
chromadb==1.0.13
sentence-transformers==4.1.0

# === ML/AI DEPENDENCIES ===
numpy==2.3.1
transformers==4.52.4
torch
tokenizers==0.21.2
scikit-learn==1.7.0
scipy==1.16.0

# === HTTP & ASYNC ===
httpx==0.28.1
aiofiles==24.1.0
requests==2.32.4

# === DOCUMENT PROCESSING ===
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5
pandas==2.3.0
chardet==5.2.0

# === UTILITIES ===
python-dateutil==2.9.0.post0

# === TESTING ===
# pytest>=8.3.0
# pytest-asyncio>=0.24.0

# === LOGGING ===
# loguru>=0.7.0

# === DEVELOPMENT ===
# black>=24.10.0
# flake8>=7.1.0

# Instalación:
# source chatbot_venv/bin/activate
# pip install -r requirements-final.txt